Move flags.hpp under hostdevcommon
